Move new configurations to SMTPConfiguration
